Rename OJMicroline to WD5API
In preparation for adding support for the WG4 series, rename `OJMicroline` to `WD5API` (since it is used for communicating with the WD5-series API).

Leave an alias in place for backwards compatibility.

The next step will be to extract a base class and add a new `WG4API` class that communicates with the WG4-series API.